Use Vector instead of List to hold Series in Plot
This commit changes the internal implementation of Plot to hold a vector of series, rather than a list. This allows us to append new series, rather than pre-pending, which means series are plotted in the correct order. It also allows for O(1) access to a random series, which is a good thing.