First commit towards Figure refactor.
The first step in allowing different subplots on the same figure is to
split out Plot behaviour (now for individual suplots) and Figure
behaviour, for the figure as a whole.

This is the first commit on the road to this split. Note that much of
the functionality is currently broken, and the tests don't compile, but
the following runs:

  val xs = Vector(1, 2, 3)
  val ys = Vector(4, 5, 6)

  val f = Figure().plot(CartesianSubplot().withScatter(xs, ys))

  import co.theasi.plotly.writer

  writer.FigureWriter.draw(f, "test-125")