Allow setting timeouts when connecting to Plotly.
Either Plotly or my network connection was particularly slow this
morning, so the tests were failing because the graph requests were
taking too long to come back from Plotly.

This commit gives the ability to configure timeouts when defining the
server, and sets a huge timeout (20s) for the tests.